imageCategorizationFaceRecognition/requirements.txt
python==3.10
click==8.1.8
cmake==4.0.0
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
dlib==19.22.99
exceptiongroup==1.2.2
face-recognition==1.3.0
face-recognition-models==0.3.0
iniconfig==2.1.0
numpy==1.22.0
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pytest==8.3.5
tomli==2.2.1
